# internal/api/gen/server.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package gen

import (
	"context"
	"encoding/json"
	"fmt"
	"io"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	strictnethttp "github.com/oapi-codegen/runtime/strictmiddleware/nethttp"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Retrieve a grayscale 28x28 PNG digit image sample by image index from 0 to 9999
	// (GET /digits)
	RetrieveDigit(w http.ResponseWriter, r *http.Request, params RetrieveDigitParams)
	// Recognize a grayscale 28x28 digit PNG image provided in the request body
	// (POST /digits/{digit}/indices/{id})
	RecognizeDigit(w http.ResponseWriter, r *http.Request, digit int8, id int)
	// Check status of training of 60000 grayscale 28x28 PNG digit image indices
	// (GET /status)
	TrainingStatus(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Retrieve a grayscale 28x28 PNG digit image sample by image index from 0 to 9999
// (GET /digits)
func (_ Unimplemented) RetrieveDigit(w http.ResponseWriter, r *http.Request, params RetrieveDigitParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Recognize a grayscale 28x28 digit PNG image provided in the request body
// (POST /digits/{digit}/indices/{id})
func (_ Unimplemented) RecognizeDigit(w http.ResponseWriter, r *http.Request, digit int8, id int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Check status of training of 60000 grayscale 28x28 PNG digit image indices
// (GET /status)
func (_ Unimplemented) TrainingStatus(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// RetrieveDigit operation middleware
func (siw *ServerInterfaceWrapper) RetrieveDigit(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params RetrieveDigitParams

	// ------------- Required query parameter "id" -------------

	if paramValue := r.URL.Query().Get("id"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "id"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "id", r.URL.Query(), &params.Id)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RetrieveDigit(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// RecognizeDigit operation middleware
func (siw *ServerInterfaceWrapper) RecognizeDigit(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "digit" -------------
	var digit int8

	err = runtime.BindStyledParameterWithOptions("simple", "digit", chi.URLParam(r, "digit"), &digit, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "digit", Err: err})
		return
	}

	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RecognizeDigit(w, r, digit, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// TrainingStatus operation middleware
func (siw *ServerInterfaceWrapper) TrainingStatus(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.TrainingStatus(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/digits", wrapper.RetrieveDigit)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/digits/{digit}/indices/{id}", wrapper.RecognizeDigit)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/status", wrapper.TrainingStatus)
	})

	return r
}

type RecognitionResponseJSONResponse Statistics

type StatusResponseJSONResponse IndexStatus

type RetrieveDigitRequestObject struct {
	Params RetrieveDigitParams
}

type RetrieveDigitResponseObject interface {
	VisitRetrieveDigitResponse(w http.ResponseWriter) error
}

type RetrieveDigit200ResponseHeaders struct {
	XDigitExpected int
}

type RetrieveDigit200ImagepngResponse struct {
	Body          io.Reader
	Headers       RetrieveDigit200ResponseHeaders
	ContentLength int64
}

func (response RetrieveDigit200ImagepngResponse) VisitRetrieveDigitResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "image/png")
	if response.ContentLength != 0 {
		w.Header().Set("Content-Length", fmt.Sprint(response.ContentLength))
	}
	w.Header().Set("X-Digit-Expected", fmt.Sprint(response.Headers.XDigitExpected))
	w.WriteHeader(200)

	if closer, ok := response.Body.(io.ReadCloser); ok {
		defer closer.Close()
	}
	_, err := io.Copy(w, response.Body)
	return err
}

type RecognizeDigitRequestObject struct {
	Digit int8 `json:"digit"`
	Id    int  `json:"id"`
	Body  io.Reader
}

type RecognizeDigitResponseObject interface {
	VisitRecognizeDigitResponse(w http.ResponseWriter) error
}

type RecognizeDigit200JSONResponse struct {
	RecognitionResponseJSONResponse
}

func (response RecognizeDigit200JSONResponse) VisitRecognizeDigitResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type TrainingStatusRequestObject struct {
}

type TrainingStatusResponseObject interface {
	VisitTrainingStatusResponse(w http.ResponseWriter) error
}

type TrainingStatus200JSONResponse struct{ StatusResponseJSONResponse }

func (response TrainingStatus200JSONResponse) VisitTrainingStatusResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Retrieve a grayscale 28x28 PNG digit image sample by image index from 0 to 9999
	// (GET /digits)
	RetrieveDigit(ctx context.Context, request RetrieveDigitRequestObject) (RetrieveDigitResponseObject, error)
	// Recognize a grayscale 28x28 digit PNG image provided in the request body
	// (POST /digits/{digit}/indices/{id})
	RecognizeDigit(ctx context.Context, request RecognizeDigitRequestObject) (RecognizeDigitResponseObject, error)
	// Check status of training of 60000 grayscale 28x28 PNG digit image indices
	// (GET /status)
	TrainingStatus(ctx context.Context, request TrainingStatusRequestObject) (TrainingStatusResponseObject, error)
}

type StrictHandlerFunc = strictnethttp.StrictHTTPHandlerFunc
type StrictMiddlewareFunc = strictnethttp.StrictHTTPMiddlewareFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// RetrieveDigit operation middleware
func (sh *strictHandler) RetrieveDigit(w http.ResponseWriter, r *http.Request, params RetrieveDigitParams) {
	var request RetrieveDigitRequestObject

	request.Params = params

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.RetrieveDigit(ctx, request.(RetrieveDigitRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RetrieveDigit")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(RetrieveDigitResponseObject); ok {
		if err := validResponse.VisitRetrieveDigitResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// RecognizeDigit operation middleware
func (sh *strictHandler) RecognizeDigit(w http.ResponseWriter, r *http.Request, digit int8, id int) {
	var request RecognizeDigitRequestObject

	request.Digit = digit
	request.Id = id

	request.Body = r.Body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.RecognizeDigit(ctx, request.(RecognizeDigitRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RecognizeDigit")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(RecognizeDigitResponseObject); ok {
		if err := validResponse.VisitRecognizeDigitResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// TrainingStatus operation middleware
func (sh *strictHandler) TrainingStatus(w http.ResponseWriter, r *http.Request) {
	var request TrainingStatusRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.TrainingStatus(ctx, request.(TrainingStatusRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "TrainingStatus")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(TrainingStatusResponseObject); ok {
		if err := validResponse.VisitTrainingStatusResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}