# internal/api/gen/types.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.1.0 DO NOT EDIT.
package gen

// IndexStatus Index Training Status
type IndexStatus struct {
	IndexCount    int `json:"indexCount"`
	IndicesStored int `json:"indicesStored"`
}

// Statistics Digit Recognition Statistics
type Statistics struct {
	CorrectGuesses int   `json:"correctGuesses"`
	DigitExpected  int8  `json:"digitExpected"`
	DigitFound     int8  `json:"digitFound"`
	DigitIndex     int   `json:"digitIndex"`
	Duration       int64 `json:"duration"`
	MaxDuration    int64 `json:"maxDuration"`
	MinDuration    int64 `json:"minDuration"`
	Success        bool  `json:"success"`
	TotalDuration  int64 `json:"totalDuration"`
	WrongGuesses   int   `json:"wrongGuesses"`
}

// RecognitionResponse Digit Recognition Statistics
type RecognitionResponse = Statistics

// StatusResponse Index Training Status
type StatusResponse = IndexStatus

// RetrieveDigitParams defines parameters for RetrieveDigit.
type RetrieveDigitParams struct {
	// Id A sample image ID from the MNIST test dataset
	Id int `form:"id" json:"id"`
}